refactor: 타임아웃 함수 분리
